database/requirements.txt
asyncio==3.4.3
asyncpg==0.30.0
greenlet==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2